Product Variations

+++ carts/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, HttpResponseRedirect
 from django.urls import reverse
 from .models import Cart, CartItem
-from products.models import Product
+from products.models import Product, Variation
 
 
 # Create your views here.
@@ -21,7 +21,7 @@ def view(request):
     return render(request, template, context)
 
 
-def update_cart(request, product_slug, qty):
+def update_cart(request, product_slug):
     request.session.set_expiry(900)
     try:
         _id = request.session['cart_id']
@@ -37,25 +37,45 @@ def update_cart(request, product_slug, qty):
         print(err)
     except Exception:
         pass
-    cart_item, created = CartItem.objects.get_or_create(cart=cart, product=product)
-    if created:
-        pass
-    if qty == 0:
-        cart_item.quantity = 0
-        cart_item.delete()
-    else:
-        cart_item.quantity = qty
-        cart_item.save()
-    # if cart_item not in cart.items.all():
-    #     cart.items.add(cart_item)
-    # else:
-    #     cart.items.remove(cart_item)
-    new_total = 0.00
-    for item in cart.cartitem_set.all():
-        brand_total = float(item.product.price) * item.quantity
-        new_total += brand_total
 
-    request.session['items_total'] = cart.cartitem_set.count()
-    cart.total = new_total
-    cart.save()
+    product_var = []
+    if request.method == 'POST':
+        qty = request.POST['qty']
+        for item in request.POST:
+            if item != 'qty':
+                key = item
+                value = request.POST.get(key)
+                print(value)
+                try:
+                    variation = Variation.objects.get(id=value)
+                    product_var.append(variation)
+                except:
+                    pass
+        print(product_var)
+
+        cart_item, created = CartItem.objects.get_or_create(cart=cart, product=product)
+        if created:
+            pass
+        if qty and int(qty) <= 0:
+            cart_item.quantity = 0
+            cart_item.delete()
+        elif qty:
+            if len(product_var) > 0:
+                cart_item.variations.clear()
+                for item in product_var:
+                    cart_item.variations.add(item)
+            cart_item.quantity = qty
+            # cart_item.notes = notes
+            cart_item.save()
+        else:
+            pass
+
+        new_total = 0.00
+        for item in cart.cartitem_set.all():
+            brand_total = float(item.product.price) * item.quantity
+            new_total += brand_total
+
+        request.session['items_total'] = cart.cartitem_set.count()
+        cart.total = new_total
+        cart.save()
     return HttpResponseRedirect(reverse("cart_view"))
